Release 6.4.3

+++ tests/modules/Import/Bug50438Test.php
@@ -89,10 +89,7 @@ class Bug50438Test extends Sugar_PHPUnit_Framework_TestCase
     public function testParentsAreRelatedDuringImport()
     {
 
-        //$file = 'upload://test50438.csv';
-        $upload_dir = isset($GLOBALS['sugar_config']['upload_dir']) ? $GLOBALS['sugar_config']['upload_dir'] : 'upload';
-        $file = $upload_dir . '/test50438.csv';
-
+        $file = 'upload://test50438.csv';
         $ret = file_put_contents($file, $this->fileArr);
         $this->assertGreaterThan(0, $ret, 'Failed to write to '.$file .' for content '.var_export($this->fileArr,true));
 
@@ -131,7 +128,12 @@ class Bug50438Test extends Sugar_PHPUnit_Framework_TestCase
         //test that the contact id is in the array of related contacts.
         $this->assertContains($this->contact->id, $related_contacts,' Contact was not related during simulated import despite being set in related parent id');
         unset($call);
-        unlink($file);
+
+        /*
+        if (is_file($file)) {
+            unlink($file);
+        }
+        */
     }
 
 }
